+++ frontend/src/core/loaders/components/CircleLoader.css
@@ -4,4 +4,3 @@
     padding: 40px 0 10px;
     text-align: center;
 }
-
