+++ frontend/src/core/loaders/components/WaveLoader.js
@@ -8,12 +8,12 @@ import './WaveLoader.css';
 import type { LocaleState } from 'modules/core/locales/reducer';
 import type { L10nState } from 'modules/core/l10n/reducer';
 
-type Props = {|
-    l10n: L10nState,
-    locales: LocaleState,
-|};
+type Props = {
+    isLoading: boolean,
+    children?: React.Node,
+};
 
-const WaveLoadingBar = () => (
+const WaveLoadingBar = (): React.Node => (
     <div id="project-load" className="overlay">
         <div className="inner">
             <div className="animation">
@@ -32,29 +32,12 @@ const WaveLoadingBar = () => (
     </div>
 );
 
-export class WaveLoader extends React.Component<Props> {
-    render() {
-        const { l10n, locales, children } = this.props;
-        const isLoading = l10n.fetching || locales.fetching;
+const WaveLoader = (props: Props): React.Node => (
+    // Render children components, because some of fetches happen after components are rendered.
+    <React.Fragment>
+        { props.isLoading ? <WaveLoadingBar /> : '' }
+        { props.children }
+    </React.Fragment>
+);
 
-        // Render children components, because some of fetches happenen after components are rendered.
-        return (
-            <React.Fragment>
-                { isLoading ? <WaveLoadingBar /> : '' }
-                { children }
-            </React.Fragment>
-        );
-    }
-}
-
-const mapStateToProps = (state: Object): Props => {
-    const { l10n, locales } = state;
-
-    return {
-        l10n,
-        locales,
-    };
-};
-
-
-export default connect(mapStateToProps)(WaveLoader);
+export default WaveLoader;
